algorhythm-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.12.0
spotipy==2.23.0
python-dotenv==1.0.1
pydantic==2.6.0
requests==2.31.0
networkx==3.2.1